Popup big refactoring
All popup views merged into one
